+++ TwitchLib.Client/TwitchClient_Notice.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using TwitchLib.Client.Events;
 using TwitchLib.Client.Exceptions;
+using TwitchLib.Client.Extensions.Internal;
 using TwitchLib.Client.Interfaces;
 using TwitchLib.Client.Models.Internal;
 using TwitchLib.Communication.Extensions;
